缺少注释

+++ src/main/java/com/manage/classInfo/mapper/MoneyInMapper.java
@@ -7,5 +7,5 @@ import com.manage.classInfo.pojo.MoneyIn;
 @Mapper
 public interface MoneyInMapper {
 	void add(MoneyIn moneyIn);
-	int allMoney();
+	double allMoney();
 }
