+++ src/main/java/com/manage/classInfo/pojo/MoneyIn.java
@@ -3,7 +3,7 @@ package com.manage.classInfo.pojo;
 public class MoneyIn {
 	private int id;
 	private String name;
-	private int count;
+	private double count;
 	private String inDate;
 	public int getId() {
 		return id;
@@ -17,10 +17,10 @@ public class MoneyIn {
 	public void setName(String name) {
 		this.name = name;
 	}
-	public int getCount() {
+	public double getCount() {
 		return count;
 	}
-	public void setCount(int count) {
+	public void setCount(double count) {
 		this.count = count;
 	}
 	public String getInDate() {
@@ -29,7 +29,7 @@ public class MoneyIn {
 	public void setInDate(String inDate) {
 		this.inDate = inDate;
 	}
-	public MoneyIn(int id, String name, int count, String inDate) {
+	public MoneyIn(int id, String name, double count, String inDate) {
 		super();
 		this.id = id;
 		this.name = name;
