+++ src/main/java/com/manage/classInfo/service/impl/MoneyInServiceImpl.java
@@ -25,7 +25,7 @@ public class MoneyInServiceImpl implements MoneyInService {
 	}
 
 	@Override
-	public int allMoney() {
+	public double allMoney() {
 		return moneyInMapper.allMoney();
 	}
